requirements.txt
ansible==8.2.0
ansible-core==2.15.2
boto3==1.28.16
botocore==1.31.16
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moto==4.1.14
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pytest==7.4.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
resolvelib==1.0.1
responses==0.23.3
s3transfer==0.6.1
six==1.16.0
types-PyYAML==6.0.12.11
urllib3==1.26.16
Werkzeug==2.3.6
xmltodict==0.13.0
